[Refactoring] Enum 모듈 분리
