+++ src/main/java/aresain/loldatastats/common/type/LaneType.java
@@ -1,4 +1,4 @@
-package aresain.loldatastats.common;
+package aresain.loldatastats.common.type;
 
 public enum LaneType {
 	TOP_LANE,
