+++ src/main/java/aresain/loldatastats/common/type/TowerType.java
@@ -1,4 +1,4 @@
-package aresain.loldatastats.common;
+package aresain.loldatastats.common.type;
 
 public enum TowerType {
 	OUTER_TURRET,
